# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Document API",
        "title": "Geolocation",
        "contact": {
            "name": "API Support",
            "url": "https://simpplify.com.br/contact",
            "email": "support@swagger.io"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/advertisement/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Cria um anúncio.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Anúncio"
                ],
                "summary": "Criar um Anúncio.",
                "parameters": [
                    {
                        "description": "Requisição de Anúncio",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/advertisement.CreateAdvertisementRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informações do Anúncio",
                        "schema": {
                            "$ref": "#/definitions/advertisement.AdvertisementResponse"
                        }
                    },
                    "400": {
                        "description": "Requisição Inválida",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro Interno do Servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/advertisement/delete/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Exclui um anúncio.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Anúncio"
                ],
                "summary": "Excluir um Anúncio.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do Anúncio",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Requisição Inválida",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro Interno do Servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/advertisement/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Recupera todos os anúncios.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Anúncio"
                ],
                "summary": "Obter Todos os Anúncios.",
                "responses": {
                    "200": {
                        "description": "Lista de Anúncios",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/advertisement.AdvertisementResponseAll"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro Interno do Servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/advertisement/route": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Atualiza a rota escolhida no anúncio.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Anúncio"
                ],
                "summary": "Atualizar Escolha de Rota do Anúncio.",
                "parameters": [
                    {
                        "description": "Requisição para escolha de rota do anúncio",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/advertisement.UpdateAdsRouteChooseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Requisição Inválida",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro Interno do Servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/advertisement/update": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Atualiza um anúncio.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Anúncio"
                ],
                "summary": "Atualizar um Anúncio.",
                "parameters": [
                    {
                        "description": "Requisição de Anúncio",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/advertisement.UpdateAdvertisementRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informações do Anúncio",
                        "schema": {
                            "$ref": "#/definitions/advertisement.AdvertisementResponse"
                        }
                    },
                    "400": {
                        "description": "Requisição Inválida",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Erro Interno do Servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/public/advertisement/list": {
            "get": {
                "description": "Recupera um anúncio pelo seu ID (público).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Anúncio"
                ],
                "summary": "Obter Anúncio por ID (Público).",
                "responses": {
                    "200": {
                        "description": "Informações do Anúncio",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/advertisement.AdvertisementResponseNoUser"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro Interno do Servidor",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "advertisement.AdvertisementResponse": {
            "type": "object",
            "properties": {
                "advance": {
                    "type": "string"
                },
                "agency": {
                    "type": "boolean"
                },
                "cargo_species": {
                    "type": "string"
                },
                "cargo_type": {
                    "type": "string"
                },
                "cargo_weight": {
                    "type": "number"
                },
                "cep_destination": {
                    "type": "string"
                },
                "cep_origin": {
                    "type": "string"
                },
                "city_destination": {
                    "type": "string"
                },
                "city_origin": {
                    "type": "string"
                },
                "complement_destination": {
                    "type": "string"
                },
                "complement_origin": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_who": {
                    "type": "string"
                },
                "delivery_date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "destination": {
                    "type": "string"
                },
                "destination_lat": {
                    "type": "number"
                },
                "destination_lng": {
                    "type": "number"
                },
                "distance": {
                    "type": "integer"
                },
                "expiration_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "neighborhood_destination": {
                    "type": "string"
                },
                "neighborhood_origin": {
                    "type": "string"
                },
                "origin": {
                    "type": "string"
                },
                "origin_lat": {
                    "type": "number"
                },
                "origin_lng": {
                    "type": "number"
                },
                "payment_type": {
                    "type": "string"
                },
                "pickup_date": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "requires_tarp": {
                    "type": "boolean"
                },
                "situation": {
                    "type": "string"
                },
                "state_destination": {
                    "type": "string"
                },
                "state_origin": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "street_destination": {
                    "type": "string"
                },
                "street_number_destination": {
                    "type": "string"
                },
                "street_number_origin": {
                    "type": "string"
                },
                "street_origin": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "toll": {
                    "type": "boolean"
                },
                "tracking": {
                    "type": "boolean"
                },
                "trailer": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_who": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "vehicles_accepted": {
                    "type": "string"
                }
            }
        },
        "advertisement.AdvertisementResponseAll": {
            "type": "object",
            "properties": {
                "active_duration": {
                    "type": "string"
                },
                "active_freight": {
                    "type": "integer"
                },
                "active_there": {
                    "type": "string"
                },
                "advance": {
                    "type": "string"
                },
                "agency": {
                    "type": "boolean"
                },
                "cargo_species": {
                    "type": "string"
                },
                "cargo_type": {
                    "type": "string"
                },
                "cargo_weight": {
                    "type": "number"
                },
                "cep_destination": {
                    "type": "string"
                },
                "cep_origin": {
                    "type": "string"
                },
                "city_destination": {
                    "type": "string"
                },
                "city_origin": {
                    "type": "string"
                },
                "complement_destination": {
                    "type": "string"
                },
                "complement_origin": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_who": {
                    "type": "string"
                },
                "delivery_date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "destination": {
                    "type": "string"
                },
                "destination_lat": {
                    "type": "number"
                },
                "destination_lng": {
                    "type": "number"
                },
                "distance": {
                    "type": "integer"
                },
                "expiration_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "neighborhood_destination": {
                    "type": "string"
                },
                "neighborhood_origin": {
                    "type": "string"
                },
                "origin": {
                    "type": "string"
                },
                "origin_lat": {
                    "type": "number"
                },
                "origin_lng": {
                    "type": "number"
                },
                "payment_type": {
                    "type": "string"
                },
                "pickup_date": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "requires_tarp": {
                    "type": "boolean"
                },
                "route_choose": {
                    "$ref": "#/definitions/new_routes.RouteOutput"
                },
                "route_index_choose": {
                    "type": "integer"
                },
                "situation": {
                    "type": "string"
                },
                "state_destination": {
                    "type": "string"
                },
                "state_origin": {
                    "type": "string"
                },
                "street_destination": {
                    "type": "string"
                },
                "street_number_destination": {
                    "type": "string"
                },
                "street_number_origin": {
                    "type": "string"
                },
                "street_origin": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "toll": {
                    "type": "boolean"
                },
                "tracking": {
                    "type": "boolean"
                },
                "trailer": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_who": {
                    "type": "string"
                },
                "user_city": {
                    "type": "string"
                },
                "user_email": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "user_name": {
                    "type": "string"
                },
                "user_phone": {
                    "type": "string"
                },
                "user_state": {
                    "type": "string"
                },
                "vehicles_accepted": {
                    "type": "string"
                }
            }
        },
...
        "advertisement.CreateAdvertisementRequest": {
            "type": "object",
            "properties": {
                "advance": {
                    "type": "string"
                },
                "agency": {
                    "type": "boolean"
                },
                "cargo_species": {
                    "type": "string"
                },
                "cargo_type": {
                    "type": "string"
                },
                "cargo_weight": {
                    "type": "number"
                },
                "cep_destination": {
                    "type": "string"
                },
                "cep_origin": {
                    "type": "string"
                },
                "city_destination": {
                    "type": "string"
                },
                "city_origin": {
                    "type": "string"
                },
                "complement_destination": {
                    "type": "string"
                },
                "complement_origin": {
                    "type": "string"
                },
                "created_who": {
                    "type": "string"
                },
                "delivery_date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "destination": {
                    "type": "string"
                },
                "distance": {
                    "type": "integer"
                },
                "expiration_date": {
                    "type": "string"
                },
                "neighborhood_destination": {
                    "type": "string"
                },
                "neighborhood_origin": {
                    "type": "string"
                },
                "origin": {
                    "type": "string"
                },
                "payment_type": {
                    "type": "string"
                },
                "pickup_date": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "requires_tarp": {
                    "type": "boolean"
                },
                "state_destination": {
                    "type": "string"
                },
                "state_origin": {
                    "type": "string"
                },
                "street_destination": {
                    "type": "string"
                },
                "street_number_destination": {
                    "type": "string"
                },
                "street_number_origin": {
                    "type": "string"
                },
                "street_origin": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "toll": {
                    "type": "boolean"
                },
                "tracking": {
                    "type": "boolean"
                },
                "trailer": {
                    "type": "string"
                },
                "vehicles_accepted": {
                    "type": "string"
                }
            }
        },
        "advertisement.UpdateAdsRouteChooseRequest": {
            "type": "object",
            "properties": {
                "advertisement_id": {
                    "type": "integer"
                },
                "new_route": {
                    "type": "integer"
                }
            }
        },
        "advertisement.UpdateAdvertisementRequest": {
            "type": "object",
            "properties": {
                "advance": {
                    "type": "string"
                },
                "agency": {
                    "type": "boolean"
                },
                "cargo_species": {
                    "type": "string"
                },
                "cargo_type": {
                    "type": "string"
                },
                "cargo_weight": {
                    "type": "number"
                },
                "cep_destination": {
                    "type": "string"
                },
                "cep_origin": {
                    "type": "string"
                },
                "city_destination": {
                    "type": "string"
                },
                "city_origin": {
                    "type": "string"
                },
                "complement_destination": {
                    "type": "string"
                },
                "complement_origin": {
                    "type": "string"
                },
                "delivery_date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "destination": {
                    "type": "string"
                },
                "destination_lat": {
                    "type": "number"
                },
                "destination_lng": {
                    "type": "number"
                },
                "distance": {
                    "type": "integer"
                },
                "expiration_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "neighborhood_destination": {
                    "type": "string"
                },
                "neighborhood_origin": {
                    "type": "string"
                },
                "origin": {
                    "type": "string"
                },
                "origin_lat": {
                    "type": "number"
                },
                "origin_lng": {
                    "type": "number"
                },
                "payment_type": {
                    "type": "string"
                },
                "pickup_date": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "requires_tarp": {
                    "type": "boolean"
                },
                "situation": {
                    "type": "string"
                },
                "state_destination": {
                    "type": "string"
                },
                "state_origin": {
                    "type": "string"
                },
                "street_destination": {
                    "type": "string"
                },
                "street_number_destination": {
                    "type": "string"
                },
                "street_number_origin": {
                    "type": "string"
                },
                "street_origin": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "toll": {
                    "type": "boolean"
                },
                "tracking": {
                    "type": "boolean"
                },
                "trailer": {
                    "type": "string"
                },
                "vehicles_accepted": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}